Reduce granularity of reset tests
Different browsers, Chrome in this case, result in different
outputs. Since these are not bugs, but rather behavioral differences
between browsers and we are not maintaining the actual rulesets (just
glamor's ability to insert them), the tests have been changed to
detect the count of rules inserted instead of string matching rule
output.