Allows setting nonce on generated style tags
Uses a feature of webpack to extract a set nonce value on the generated
style tags to allow users to maintain a Content Security Policy that
does not need to use `unsafe-inline` for the styles.

Updates the `makeStyleTag` function to optionally add the `nonce`
attribute if the `__webpack_nonce__` value is set.

Adds documentation of the feature at docs/nonce.md