Use current version of inline-style-prefixer
This crosses the v5.0.0 version, which required some changes to the
import paths. All tests still pass.

It turns out crossing this threshold is important for us, because we
depend on libraries that depend on post-5.0.0 versions of
inline-style-prefixer, and the two versions are incompatible. Node
resolves this by allowing multiple versions of a single package in
the same project. But we're running this code in browsers, and we're
constrained to using a bundler that also doesn't support multiple
versions. So if we can bump this library to use a post-5.0.0
version, all uses will be compatible.